Misc cleanups as followup to flex tree schema removal
