--- packages/dds/tree/src/feature-libraries/typed-schema/README.md
@@ -1,6 +1,0 @@
-# Typed Schema
-
-Schema building API which captures type information in both runtime data and TypeScript types.
-
-TODO:
-The portion of this that is not dependent on specific field kinds should be separated from the part that does depend on specific field kinds.
