+++ packages/dds/tree/src/simple-tree/api/schemaFactoryRecursive.ts
@@ -3,7 +3,6 @@
  * Licensed under the MIT License.
  */
 
-import type { Unenforced } from "../../feature-libraries/index.js";
 import type { RestrictiveStringRecord } from "../../util/index.js";
 import type { InsertableObjectFromSchemaRecord } from "../objectNode.js";
 
@@ -22,7 +21,7 @@ import type {
 	WithType,
 	TreeNode,
 } from "../core/index.js";
-import type { FieldSchemaUnsafe } from "../typesUnsafe.js";
+import type { FieldSchemaUnsafe, Unenforced } from "./typesUnsafe.js";
 
 export function createFieldSchemaUnsafe<
 	Kind extends FieldKind,
