+++ packages/dds/tree/src/test/simple-tree/typesUnsafe.spec.ts
@@ -4,7 +4,7 @@
  */
 
 // eslint-disable-next-line import/no-internal-modules
-import type { ReadonlyMapInlined } from "../../simple-tree/typesUnsafe.js";
+import type { ReadonlyMapInlined } from "../../simple-tree/api/typesUnsafe.js";
 // eslint-disable-next-line import/no-internal-modules
 import type { numberSchema } from "../../simple-tree/leafNodeSchema.js";
 import type { areSafelyAssignable, requireTrue } from "../../util/index.js";
